Feature: Analytics on website traffic over time
we show a line chart that shows the total number of visits to the website per week for a given period of time. The user specifies a period of time using the bootstrap datepicker.
we show the number of times events have happened each week between the start dates and end dates.

For example, if the start date is 2/1/15 and the end date  is 2/17/15, the chart contains the number of times events have occurred from:
2/1/15 - 2/6/15 (week1)
2/6/15 - 2/12/15 (week 2)
2/13/15 - 2/17/15 (week 3 - incomplete week)